tests/wpt/web-platform-tests/element-timing/observe-child-element.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Element Timing: observe elements from same-origin iframes</title>
<body>
<style>
body {
  margin: 0;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/element-timing-helpers.js"></script>
<script>
  let beforeRender;
  async_test((t) => {
    const observer = new PerformanceObserver(
      t.step_func_done((entryList) => {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        checkElement(entry, 'my_image', beforeRender);
        // Assume viewport has size at least 100, so the element is fully visible.
        checkRect(entry, [0, 100, 0, 100]);
      })
    );
    observer.observe({entryTypes: ['element']});
    // We add the iframe during onload to be sure that the observer is registered
    // in time for it to observe the element timing.
    // TODO(npm): change observer to use buffered flag.
    window.onload = () => {
      // Add iframe with an image of width and height equal to 100.
      const iframe = document.createElement('iframe');
      iframe.src = 'resources/iframe-with-square.html';
      document.body.appendChild(iframe);
      beforeRender = performance.now();
    };
  }, 'Element from same-origin iframe is observable.');
</script>

</body>
